turned put from promise to async await

+++ server/routes/api/product-routes.js
@@ -1,4 +1,5 @@
 const router = require('express').Router();
+const { response } = require('express');
 const { Product, Category, Tag, ProductTag } = require('../../models');
 
 // The `/api/products` endpoint
@@ -6,7 +7,7 @@ const { Product, Category, Tag, ProductTag } = require('../../models');
 // get all products
 router.get('/', async (req, res) => {
   try {
-    const allProducts = await Prloduct.findAll({
+    const allProducts = await Product.findAll({
       include: [{
                   model: Category,
                   as: 'category'
@@ -63,7 +64,7 @@ router.post('/', async (req, res) => {
       stock,
     })
 
-    if (tagIds.length) {
+    if (tagIds && tagIds.length) {
       try {
         const productTagIdArr = req.body.tagIds.map((tag_id) => {
           return {
@@ -91,45 +92,50 @@ router.post('/', async (req, res) => {
 });
 
 // update product
-router.put('/:id', (req, res) => {
-  // update product data
-  Product.update(req.body, {
-    where: {
-      id: req.params.id,
-    },
-  })
-    .then((product) => {
-      // find all associated tags from ProductTag
-      return ProductTag.findAll({ where: { product_id: req.params.id } });
+router.put('/:id', async (req, res) => {
+  try {
+      // update product data
+    await Product.update(req.body, {
+      where: {
+        id: req.params.id,
+      },
     })
-    .then((productTags) => {
-      // get list of current tag_ids
-      const productTagIds = productTags.map(({ tag_id }) => tag_id);
-      // create filtered list of new tag_ids
-      const newProductTags = req.body.tagIds
-        .filter((tag_id) => !productTagIds.includes(tag_id))
-        .map((tag_id) => {
-          return {
-            product_id: req.params.id,
-            tag_id,
-          };
-        });
-      // figure out which ones to remove
-      const productTagsToRemove = productTags
-        .filter(({ tag_id }) => !req.body.tagIds.includes(tag_id))
-        .map(({ id }) => id);
 
-      // run both actions
-      return Promise.all([
-        ProductTag.destroy({ where: { id: productTagsToRemove } }),
-        ProductTag.bulkCreate(newProductTags),
-      ]);
-    })
-    .then((updatedProductTags) => res.json(updatedProductTags))
-    .catch((err) => {
-      // console.log(err);
-      res.status(400).json(err);
-    });
+    if (req.body.tag_id) {
+      try {
+        const productTags = await ProductTag.findAll({ where: { product_id: req.params.id } });
+        const productTagsArray = productTags.map(({ tag_id }) => tag_id)
+  
+        const newProductTags = req.body.tag_id
+          .filter((tag_id) => !productTagsArray.includes(tag_id))
+            .map((tag_id) => {
+              return {
+                product_id: req.params.id,
+                tag_id,
+              };
+            });
+    
+        const productTagsToRemove = productTags
+        .filter(({ tag_id }) => !req.body.tag_id.includes(tag_id))
+        .map(({ id }) => id);
+    
+        await ProductTag.destroy({ where: { id: productTagsToRemove } }),
+        await ProductTag.bulkCreate(newProductTags),
+        res.status(200).json({message: "Product has been successfully updated"})
+      } catch (error) {
+        console.log(error)
+        res.status(500).json({error: "Sorry we couldn't update your product tags at this time."})
+      }
+    }
+    
+    res.status(200).json({message: "Product has been successfully updated"})
+
+  } catch (error) {
+    console.log(error)
+    res.status(500).json({error: "Sorry we couldn't update your product at this time."})
+  }
+  
+
 });
 
 // delete one product by its `id` value
@@ -143,7 +149,6 @@ router.delete('/:id', async (req, res) => {
       message: "The category has been successfully deleted",
     })
   } catch (error) {
-    console.log(error)
     res.status(500).json({
       message: "We were unable to delete the product at this time. Please try again later."
     })
